backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
faker
pytest
pytest-asyncio
httpx
pytest-cov
python-dotenv 